ragas-eval-final/pyproject.toml
[project]
name = "ragas-eval-final"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.6",
    "google-ai-haystack>=4.0.1",
    "grpcio==1.60.1",
    "haystack-ai>=2.8.0",
    "langchain-groq>=0.2.3",
    "langchain-openai>=0.3.0",
    "opensearch-haystack>=1.2.0",
    "opensearch-py>=2.8.0",
    "python-dotenv>=1.0.1",
    "ragas==0.2.11",
    "sacrebleu>=2.5.1",
    "sentence-transformers>=3.3.1",
    "uvicorn>=0.34.0",
]
